feat!: complete rewrite code and output
