+++ .changeset/nine-bikes-refuse.md
@@ -1,0 +1,5 @@
+---
+'@mheob/changeset-changelog': major
+---
+
+complete rewrite of the code and output
